delete with policy authoriaztion

+++ app/Http/Controllers/PostController.php
@@ -53,9 +53,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        if (!$post->ownedBy(auth()->user())) {
-            dd("no");
-        }
+        //authrization
+        $this->authorize('delete', $post);
 
         $post->delete();
         return back();
